fix bug call api facemath

+++ ekyc/src/main/java/ai/ftech/ekyc/data/source/remote/model/CaptureFaceData.kt
@@ -22,7 +22,7 @@ class CaptureFaceData {
 
     @SerializedName("matchPercent")
     @Expose
-    var matchPercent: Int? = null
+    var matchPercent: Double? = null
 
     @SerializedName("imageUrl")
     @Expose
